refactor(categories): move CategoryBottomSheet to design_system component package for better organization
